fix from comments revert golden files
