+++ examples/booking_app/lib/lib_router/routes/routes.dart
@@ -1,4 +1,4 @@
-part of '../../router.dart';
+part of '../router.dart';
 
 @TypedGoRoute<HomeRoutes>(
   path: RoutesPath.home,
@@ -22,6 +22,31 @@ class HomeRoutes extends GoRouteData implements RouteData {
           navigationType: type,
         ),
       );
+
+  @override
+  String get routeLocation => location;
+}
+
+@immutable
+class HotelDetailsRoutes extends GoRouteData implements RouteData {
+  const HotelDetailsRoutes(
+    this.type,
+    this.id,
+  );
+
+  final NavigationItemType type;
+  final String id;
+
+  @override
+  Page<Function> buildPage(BuildContext context, GoRouterState state) {
+    return MaterialPage(
+      key: state.pageKey,
+      child: HotelDetailsWithDependencies(
+        hotelId: id,
+        hotel: state.extra as Hotel?,
+      ),
+    );
+  }
 
   @override
   String get routeLocation => location;
